requirements.txt
torch
transformers
peft
datasets
accelerate
bitsandbytes
fastapi
uvicorn
pandas
pyarrow
scikit-learn
pyyaml
numpy
pydantic
